@@ -2,7 +2,7 @@ import { ChangeDetectionStrategy, ChangeDetectorRef, Component, Inject, OnInit }
 import { Router } from '@angular/router';
 import { PromptComponent, IServerResponse, PromptCallbackType } from '@autoinquirer/shared';
 import { AuthService } from '../services/auth.service';
-import { FormGroup } from '@angular/forms';
+import { UntypedFormGroup } from '@angular/forms';
 import { FormlyJsonschema } from '@ngx-formly/core/json-schema';
 import { FormlyFieldConfig } from '@ngx-formly/core';
 import { startWith, map, tap } from 'rxjs/operators';
@@ -50,7 +50,7 @@ export class AutoinquirerAuthRegisterComponent implements PromptComponent, OnIni
   prompt: any = {};
   user: NewUser = { email: '', password: '', passwordConfirm: ''};
   loading = false;
-  form = new FormGroup({});
+  form = new UntypedFormGroup({});
   fields: FormlyFieldConfig[] = [];
   callback: PromptCallbackType;
   
@@ -62,7 +62,7 @@ export class AutoinquirerAuthRegisterComponent implements PromptComponent, OnIni
       private cdref: ChangeDetectorRef
     ) { }
 
-  fieldMap(form: FormGroup) {
+  fieldMap(form: UntypedFormGroup) {
     return (mappedField: FormlyFieldConfig, mapSource: any) => {
       if (!mappedField.templateOptions) { mappedField.templateOptions = {}; }
       mappedField.templateOptions.disabled = mappedField.templateOptions.disabled || mapSource.readOnly;
